+++ packages/api-reference/src/features/ApiClientModal/index.ts
@@ -1,2 +1,2 @@
 export { default as ApiClientModal } from './ApiClientModal.vue'
-export * from './useApiClient'
+export * from './hooks/useApiClient'
